Add support for imshow()
This PR adds support for the imshow() function to matplotlibcpp. imshow() displays an image in the current figure and allows for setting extra parameters such as the colormap used.

I've also added a version of the function which takes an OpenCV matrix as an input (in a separate commit), which provides a nicer interface. This requires OpenCV, but I've made this functionality optional: you have to explicitly define WITH_OPENCV if you want it.